Revert "More pointer casting fixes"
This reverts commit 56ac4e60202c45caaca4238119e2975dbb3b56a2.

This is wrong.

+++ neo/d3xp/ai/AAS_routing.cpp
@@ -167,7 +167,7 @@ void idAASLocal::CalculateAreaTravelTimes(void) {
 		}
 	}
 
-	assert( ( (size_t) bytePtr - (size_t) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
+	assert( ( (unsigned int) bytePtr - (unsigned int) areaTravelTimes ) <= numAreaTravelTimes * sizeof( unsigned short ) );
 }
 
 /*
